New package openssl-certs
The Public Key Inrastructure is used for many security issues
in a Linux system. In order for a certificate to be trusted,
it must be signed by a trusted agent called a Certificate Authority (CA).
The certificates loaded by this section are from the list on the
Mozilla version control system and formats it into a form used by
OpenSSL-1.0.0e. The certificates can also be used by other
applications either directly of indirectly through openssl.