New package subvesion
The goal of the Subversion project is to build a version control
system that is a compelling replacement for CVS in the open source
community.  The software is released under an Apache/BSD-style open
source license.

This is a meta-package.  It installs no files, but it depends on all
the Subversion components, including the base package with the
subversion client and server, the apache module, and four language
bindings.  (If you just want the basic svn and svnadmin programs see
devel/subversion-base.)