New package tar
tar - The GNU version of the tar archiving utility
The Tar program provides the ability to create tar
archives, as well as various other kinds of manipulation.
For example, you can use Tar on previously created
archives to extract files, to store additional files, or
to update or list files which were already stored.